refactor: split commands module
